requirements.txt
flask==3.1.0
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-jwt-extended==4.7.1
flask-migrate==4.0.4
flask-socketio==5.5.1
flask-limiter==3.5.0
flask-talisman==1.1.0
gunicorn==20.1.0
numpy<2.0.0
pandas==1.5.3
scikit-learn==1.2.2
tensorflow==2.19.0
opencv-python==4.7.0.72
pillow==11.1.0
mediapipe==0.9.1.0
pydantic==1.10.6
pytest==7.3.1
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.5
marshmallow==3.19.0
celery==5.2.7
redis==4.5.4
bcrypt==4.0.1
email-validator==2.0.0
itsdangerous==2.2.0
python-dateutil==2.8.2
eventlet==0.33.3 